Version 1.2.1
Commencing a major overhaul of the code base.

I am creating LESS files to manage my CSS and the new /css/main.css file will be my development CSS file. It will be minified in /css/min/main.min.css when it is ready to go live.

I will be using a 12-grid responsive layout for sections of the code to modernise the CSS to some of the latest techniques I have learnt.

The CSS and Javascript folders have been restructured. I will also minify the Javascript files in the future.

Changes

- Reduced transition of menu items.
- Added a new break point for desktops with over 1400 pixel width.
- Modified fonts, body and footer CSS code.